requirements.txt
flask==2.0.3
semver==2.13.0
psycopg2-binary==2.9.3
